utils/pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-discoveryengine>=0.16.0",
]
